// app/controller/enums.ts
export enum PlayerActions {
    STANCE = "STANCE",
    MOVE_LEFT = "MOVE_LEFT",
    MOVE_RIGHT = "MOVE_RIGHT",
    JUMP = "JUMP",
    CROUCH = "CROUCH",
    SHOOT = "SHOOT"
}

export enum WorldItems {
    GROUND = "GROUND",
    BRICK_1 = "BRICK_1",
    BRICK_2 = "BRICK_2",
    BRICK_3 = "BRICK_3",
    QUESTION_MARK_1 = "QUESTION_MARK_1",
    QUESTION_MARK_2 = "QUESTION_MARK_2",
    QUESTION_MARK_3 = "QUESTION_MARK_3",
    QUESTION_MARK_4 = "QUESTION_MARK_4",
    PIPE_1 = "PIPE_1",
    PIPE_2 = "PIPE_2",
    PIPE_3 = "PIPE_3",
}

export enum WorldPhysics {
    MAKE_GRAVITY = "MAKE_GRAVITY"
}

export enum ControlsButtons {
    BUTTON_LEFT = "BUTTON_LEFT",
    BUTTON_RIGHT = "BUTTON_RIGHT",
    BUTTON_JUMP = "BUTTON_JUMP",
    BUTTON_GUN = "BUTTON_GUN",
    BUTTON_RUN ="BUTTON_RUN",
    BUTTON_CROUCH = "BUTTON_CROUCH"
}